refactor(drivers/wps): remove resolveCache
